requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib
requests
python-dotenv
pydantic-sqlalchemy
pyjwt
python-multipart
python-jose
xrpl-py